refactor: extract compressed backend connections
